requirements.txt
accelerate==1.8.1
annotated-types==0.7.0
anyio==4.9.0
Authlib==1.6.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cmudict==1.0.33
cryptography==45.0.5
cssutils==2.11.1
distro==1.9.0
dnspython==2.7.0
EditorConfig==0.17.1
email_validator==2.2.0
filelock==3.13.1
Flask==3.1.1
Flask-Dance==7.1.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fsspec==2024.6.1
greenlet==3.2.3
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.2
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsbeautifier==1.15.4
lxml==6.0.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.3
numpy==2.3.1
oauthlib==3.3.1
openai==1.93.0
packaging==25.0
pillow==11.3.0
psutil==7.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyMuPDF==1.26.3
PyMySQL==1.1.1
PyPDF2==3.0.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-whois==0.9.5
PyYAML==6.0.2
razorpay==1.4.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rjsmin==1.2.4
safetensors==0.5.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
stripe==12.3.0
sympy==1.13.3
textstat==0.7.7
tinycss2==1.4.0
tokenizers==0.21.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.5.0
URLObject==2.4.3
user-agents==2.2.0
webencodings==0.5.1
Werkzeug==3.1.3
WTForms==3.2.1
zipp==3.23.0
